requirements.txt
redis
pyfiglet
requests
jpype1
jellyfish
